update with code review suggestions
